fixes #7, renamed uint to size_t

+++ src/linearsolver/ls_arma.h
@@ -7,13 +7,13 @@
 template<class Tx, class Tad>
 class LinearSolverArma{
  public:
-	uint n; //!< Size of the linear system, square matrix is assumed.
+	size_t n; //!< Size of the linear system, square matrix is assumed.
 	arma::Mat<Tx> rhs; //!< Matrix container for the right hand side
 	arma::Mat<Tx> dq; //!< Matrix container for the solution
 	arma::SpMat<Tx> lhs; //!< Matrix container for the left hand side
 
-	uint number_lhs_update = 0; //!< Number of times the left hand side is updated.
-	uint number_rhs_update = 0; //!< Number of times the right hand side is updated.
+	size_t number_lhs_update = 0; //!< Number of times the left hand side is updated.
+	size_t number_rhs_update = 0; //!< Number of times the right hand side is updated.
 
 	LinearSolverArma(std::shared_ptr<Mesh<Tx, Tad>> val_mesh, std::shared_ptr<Config<Tx>> val_config){
 		const auto nic = val_mesh->nic;
@@ -29,7 +29,7 @@ class LinearSolverArma{
 	void set_lhs(int nnz, unsigned int *rind, unsigned int *cind, Tx *values){
 		number_lhs_update += 1;
 		Tx value_tmp;
-		for(uint i=0; i<nnz; i++){
+		for(size_t i=0; i<nnz; i++){
 			value_tmp = values[i];
 			lhs(rind[i],cind[i]) = value_tmp;
 		}
@@ -37,7 +37,7 @@ class LinearSolverArma{
 
 	void set_rhs(Tx *val_rhs){
 		number_rhs_update += 1;
-		for(uint i=0; i<n; i++){
+		for(size_t i=0; i<n; i++){
 			rhs(i, 0) = val_rhs[i];
 		}
 	};
@@ -50,7 +50,7 @@ class LinearSolverArma{
 	};
 	
 	void solve_and_update(Tx *q, Tx under_relaxation){
-		for(int i=0; i<n; i++){
+		for(size_t i=0; i<n; i++){
 			q[i] = q[i] + dq(i,0)*under_relaxation;
 		}
 	};
