+++ src/linearsolver/ls_eigen.h
@@ -6,15 +6,15 @@
 template <class Tx, class Tad>
 class LinearSolverEigen{
  private:
-	uint n; //!< Size of the linear system, square matrix is assumed.
+	size_t n; //!< Size of the linear system, square matrix is assumed.
 	Eigen::MatrixXd rhs; //!< Matrix container for the right hand side
 	Eigen::MatrixXd dq; //!< Matrix container for the solution
 	Eigen::SparseMatrix<Tx,  Eigen::ColMajor> lhs; //!< Matrix container for the left hand side
 	std::unique_ptr<Eigen::SparseLU<Eigen::SparseMatrix<Tx>>> solver; //!< Pointer to the linear solver
 
 	bool pattern_analyzed = false; //!< To ensure that pattern is analyzed when the solver is called for the first time 
-	uint number_lhs_update = 0; //!< Number of times the left hand side is updated.
-	uint number_rhs_update = 0; //!< Number of times the right hand side is updated.
+	size_t number_lhs_update = 0; //!< Number of times the left hand side is updated.
+	size_t number_rhs_update = 0; //!< Number of times the right hand side is updated.
  public:
 	LinearSolverEigen(std::shared_ptr<Mesh<Tx, Tad>> val_mesh, std::shared_ptr<Config<Tx>> val_config){
 		const auto nic = val_mesh->nic;
@@ -40,7 +40,7 @@ class LinearSolverEigen{
 	void set_lhs(int nnz, unsigned int *rind, unsigned int *cind, Tx *values){
 		number_lhs_update += 1;
 		Tx value_tmp;
-		for(uint i=0; i<nnz; i++){
+		for(size_t i=0; i<nnz; i++){
 			value_tmp = values[i];
 			lhs.coeffRef(rind[i],cind[i]) = value_tmp;
 		}
@@ -48,7 +48,7 @@ class LinearSolverEigen{
 
 	void set_rhs(Tx *val_rhs){
 		number_rhs_update += 1;
-		for(uint i=0; i<n; i++){
+		for(size_t i=0; i<n; i++){
 			rhs(i, 0) = val_rhs[i];
 		}
 	};
@@ -70,7 +70,7 @@ class LinearSolverEigen{
 	
 	void solve_and_update(Tx *q, Tx under_relaxation){
 		solve();
-		for(int i=0; i<n; i++){
+		for(size_t i=0; i<n; i++){
 			q[i] = q[i] + dq(i,0)*under_relaxation;
 		}
 	};
